+++ components/restreamer/client/src/components/AppMix.svelte
@@ -3,10 +3,15 @@
   import { setClient, subscribe } from 'svelte-apollo';
   import Shell from './common/Shell.svelte';
   import Output from './Output.svelte';
-  import { Output as Mix, TuneVolume, TuneDelay } from '../../api/mix.graphql';
+  import {
+    Output as Mix,
+    TuneVolume,
+    TuneDelay,
+    TuneSidechain,
+  } from '../../api/mix.graphql';
   import YoutubePlayer from './common/YoutubePlayer.svelte';
 
-  const mutations = { TuneVolume, TuneDelay };
+  const mutations = { TuneVolume, TuneDelay, TuneSidechain };
 
   const gqlClient = createGraphQlClient(
     '/api-mix',