+++ components/restreamer/client/src/components/Mixin.svelte
@@ -8,21 +8,33 @@
   export let restream_id;
   export let output_id;
   export let mutations;
+  export let activeSidechainId;
 
   const tuneDelayMutation = mutation(mutations.TuneDelay);
+  const tuneSidechainMutation = mutation(mutations.TuneSidechain);
 
   let delay = 0;
+  let sidechain = false;
+  let isSidechainDisabled = false;
+
   $: {
     // Trigger Svelte reactivity watching.
     value.delay = value.delay;
-    // Move `volume` and `delay` to a separate function to omit triggering this
+    value.sidechain = value.sidechain;
+    isSidechainDisabled = !!activeSidechainId && activeSidechainId !== value.id;
+    // Move `sidechain` and `delay` to a separate function to omit triggering this
     // block when they are changed, as we're only interested in `value` changes
     // here.
     update_delay();
+    update_sidechain();
   }
 
   function update_delay() {
     delay = value.delay / 1000;
+  }
+
+  function update_sidechain() {
+    sidechain = value.sidechain;
   }
 
   async function tuneDelay() {
@@ -34,6 +46,20 @@
     };
     try {
       await tuneDelayMutation({ variables });
+    } catch (e) {
+      showError(e.message);
+    }
+  }
+
+  async function tuneSidechain() {
+    const variables = {
+      restream_id,
+      output_id,
+      mixin_id: value.id,
+      sidechain: sidechain,
+    };
+    try {
+      await tuneSidechainMutation({ variables });
     } catch (e) {
       showError(e.message);
     }
@@ -61,9 +87,11 @@
       {mutations}
       max={value.src.startsWith('ts://') ? 1000 : 200}
       mixin_id={value.id}
+      title="Mixed audio"
     />
-    <div class="delay">
+    <div class="mixin-options">
       <i class="far fa-clock" title="Delay" />
+      <span>Delay</span>
       <input
         class="uk-input"
         type="number"
@@ -71,29 +99,50 @@
         step="0.1"
         bind:value={delay}
         on:change={tuneDelay}
+        title="Delay"
       />
       <span>s</span>
+      <i class="fas fa-link" title="Sidechain" />
+      <span>Sidechain</span>
+      <input
+        class="uk-checkbox"
+        type="checkbox"
+        bind:checked={sidechain}
+        disabled={isSidechainDisabled}
+        on:change={tuneSidechain}
+        title="Sidechain"
+      />
     </div>
   </div>
 </template>
 
 <style lang="stylus">
-  .fa-wave-square, .fa-clock
+  .fa-wave-square, .fa-clock, .fa-link
     font-size: 10px
     color: #d9d9d9
 
   .mixin
     margin-top: 6px
 
-  .delay
+  .mixin-options
     padding-left: 17px
     font-size: 10px
 
     .uk-input
       height: auto
-      width: 40px
+      width: 35px
       padding: 0
       border: none
+      margin-top: -2px
+      text-align: right
+
+    .fa-link
+      margin-left: 15px
+
+    .uk-checkbox
+      height: 10px
+      width: 10px
+      padding: 0
       margin-top: -2px
       text-align: right
 </style>