+++ components/restreamer/src/api/graphql/mix.rs
@@ -60,6 +60,22 @@ impl MutationsRoot {
             .state()
             .tune_delay(restream_id, output_id, mixin_id, delay)
     }
+
+    /// Tunes a the specified [`Mixin.sidechain`] in this [`State`]
+    fn tune_sidechain(
+        restream_id: RestreamId,
+        output_id: OutputId,
+        mixin_id: MixinId,
+        sidechain: bool,
+        context: &Context,
+    ) -> Option<bool> {
+        context.state().tune_sidechain(
+            restream_id,
+            output_id,
+            mixin_id,
+            sidechain,
+        )
+    }
 }
 
 /// Root of all [GraphQL queries][1] in the [`Schema`].